EY-5256 forenkle resultat av forbehandling
