+++ apps/etterlatte-saksbehandling-ui/client/src/components/etteroppgjoer/components/resultatAvForbehandling/ResultatAvForbehandling.tsx
@@ -1,106 +1,55 @@
-import { Box, Heading, HelpText, HStack, Table, VStack } from '@navikt/ds-react'
-import { NOK } from '~utils/formatering/formatering'
 import { useEtteroppgjoer } from '~store/reducers/EtteroppgjoerReducer'
+import { BodyShort, Box, HStack, Label, VStack } from '@navikt/ds-react'
+import { NOK } from '~utils/formatering/formatering'
+import { EtteroppgjoerResultatType } from '~shared/types/EtteroppgjoerForbehandling'
 
 export const ResultatAvForbehandling = () => {
-  const etteroppgjoer = useEtteroppgjoer()
+  const { beregnetEtteroppgjoerResultat, behandling } = useEtteroppgjoer()
 
-  if (!etteroppgjoer || !etteroppgjoer.beregnetEtteroppgjoerResultat) {
-    return null
+  if (!beregnetEtteroppgjoerResultat) return null
+
+  const resultatTekst: Record<EtteroppgjoerResultatType, string> = {
+    TILBAKEKREVING: 'Tilbakekreving',
+    ETTERBETALING: 'Etterbetaling',
+    IKKE_ETTEROPPGJOER: 'Ikke etteroppgjør',
   }
 
-  const resultat = etteroppgjoer?.beregnetEtteroppgjoerResultat
-
   return (
-    <VStack gap="4">
-      <Heading size="large">
-        <HStack gap="2">Resultat</HStack>
-      </Heading>
-      <Box maxWidth="25rem">
-        <Table>
-          <Table.Header>
-            <Table.Row>
-              <Table.HeaderCell scope="col">Utregning</Table.HeaderCell>
-              <Table.HeaderCell scope="col">
-                <HStack justify="center">Beløp</HStack>
-              </Table.HeaderCell>
-            </Table.Row>
-          </Table.Header>
-          <Table.Body>
-            <Table.Row>
-              <Table.HeaderCell scope="row">Ny brutto stønad</Table.HeaderCell>
-              <Table.DataCell>
-                <HStack justify="end">{NOK(resultat.nyBruttoStoenad)}</HStack>
-              </Table.DataCell>
-            </Table.Row>
-            <Table.Row>
-              <Table.HeaderCell scope="row">Brutto utbetalt stønad</Table.HeaderCell>
-              <Table.DataCell>
-                <HStack justify="end">{NOK(resultat.utbetaltStoenad)}</HStack>
-              </Table.DataCell>
-            </Table.Row>
-            <Table.Row>
-              <Table.HeaderCell scope="row">
-                <HStack gap="2">
-                  Avviksbeløp +/-
-                  <HelpText>
-                    {resultat.differanse > 0
-                      ? 'Avviksbeløpet viser at det er utbetalt for mye.'
-                      : resultat.differanse < 0
-                        ? 'Avviksbeløpet viser at det er utbetalt for lite.'
-                        : 'Avviksbeløpet viser at utbetalingen har vært korrekt.'}
-                  </HelpText>
-                </HStack>
-              </Table.HeaderCell>
-              <Table.DataCell>
-                <HStack justify="end">
-                  {resultat.differanse > 0 && '+'}
-                  {NOK(resultat.differanse)}
-                </HStack>
-              </Table.DataCell>
-            </Table.Row>
-            <Table.Row>
-              <Table.HeaderCell scope="row">
-                <HStack gap="2">
-                  Toleransegrense
-                  <HelpText>
-                    Etteroppgjør skal unnlates hvis for lite utbetalt er mindre enn 25 prosent av rettsgebyret, eller
-                    hvis for mye utbetalt er mindre enn ett rettsgebyr. Jf. forskriftens § 9.
-                  </HelpText>
-                </HStack>
-              </Table.HeaderCell>
-              <Table.DataCell>
-                <HStack justify="end">
-                  {resultat.differanse > 0 ? NOK(resultat.grense.tilbakekreving) : NOK(resultat.grense.etterbetaling)}
-                </HStack>
-              </Table.DataCell>
-            </Table.Row>
-          </Table.Body>
-        </Table>
-
-        <p>
-          {(() => {
-            const diff = resultat.differanse
-            const diffBeloep = NOK(Math.abs(diff))
-
-            if (diff > 0) {
-              if (diff <= resultat.grense.tilbakekreving) {
-                return `Resultatet viser at det er utbetalt ${diffBeloep} for mye stønad i ${etteroppgjoer.behandling.aar}, men beløpet er innenfor toleransegrense for tilbakekreving, og det kreves derfor ikke tilbake.`
+    <Box
+      marginBlock="8 0"
+      paddingInline="6"
+      paddingBlock="8"
+      background="surface-action-subtle"
+      borderColor="border-action"
+      borderWidth="0 0 0 4"
+      maxWidth="42.5rem"
+    >
+      <HStack gap="2" maxWidth="fit-content">
+        <VStack gap="2" maxWidth="42.5rem" marginBlock="05 0">
+          <Label>{resultatTekst[beregnetEtteroppgjoerResultat.resultatType]}</Label>
+          <BodyShort>
+            {(() => {
+              const diff = beregnetEtteroppgjoerResultat.differanse
+              const diffBeloep = NOK(Math.abs(diff))
+              if (diff > 0) {
+                if (diff <= beregnetEtteroppgjoerResultat.grense.tilbakekreving) {
+                  return `Resultatet viser at det er utbetalt ${diffBeloep} for mye stønad i ${behandling.aar}, men beløpet er innenfor toleransegrense for tilbakekreving, og det kreves derfor ikke tilbake.`
+                }
+                return `Resultatet viser at det er utbetalt ${diffBeloep} for mye stønad i ${behandling.aar}. Beløpet blir derfor krevd tilbake.`
               }
-              return `Resultatet viser at det er utbetalt ${diffBeloep} for mye stønad i ${etteroppgjoer.behandling.aar}. Beløpet blir derfor krevd tilbake.`
-            }
 
-            if (diff < 0) {
-              if (diff >= -resultat.grense.etterbetaling) {
-                return `Resultatet viser at det er utbetalt ${diffBeloep} for lite stønad i${etteroppgjoer.behandling.aar}, men beløpet er innenfor toleransegrense for etterbetaling, og det blir derfor ikke utbetalt.`
+              if (diff < 0) {
+                if (diff >= -beregnetEtteroppgjoerResultat.grense.etterbetaling) {
+                  return `Resultatet viser at det er utbetalt ${diffBeloep} for lite stønad i${behandling.aar}, men beløpet er innenfor toleransegrense for etterbetaling, og det blir derfor ikke utbetalt.`
+                }
+                return `Resultatet viser at det er utbetalt ${diffBeloep} for lite stønad i ${behandling.aar}. Beløpet blir derfor etterbetalt.`
               }
-              return `Resultatet viser at det er utbetalt ${diffBeloep} for lite stønad i ${etteroppgjoer.behandling.aar}. Beløpet blir derfor etterbetalt.`
-            }
 
-            return `Resultatet viser ingen endring, bruker fikk utbetalt rett stønad i ${etteroppgjoer.behandling.aar}.`
-          })()}
-        </p>
-      </Box>
-    </VStack>
+              return `Resultatet viser ingen endring, bruker fikk utbetalt rett stønad i ${behandling.aar}.`
+            })()}
+          </BodyShort>
+        </VStack>
+      </HStack>
+    </Box>
   )
 }
